requirements.txt
absl-py==2.4.0
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
glfw==2.10.0
kiwisolver==1.4.9
matplotlib==3.10.8
mediapipe==0.10.32
mss==10.1.0
numpy==2.2.6
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.0
pycparser==3.0
pygame==2.6.1
PyOpenGL==3.1.10
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sounddevice==0.5.5
trimesh==4.11.2